refactor: repositoryの実装をクラスにまとめた
