+++ src/controller/users/users.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from "express";
 import { UsersUseCase } from "../../service/users/main.js";
-import { PrismaUsersRepository } from "../../prisma/users.js";
+import { PrismaUsersRepository } from "../../repository/prisma/users.js";
 
 export class UsersController {
   private _usersUsecase: UsersUseCase;
@@ -9,19 +9,19 @@ export class UsersController {
     this._usersUsecase = new UsersUseCase(repo);
   }
 
-  async getAllUsers(_req: Request, res: Response) {
+  public getAllUsers = async (_req: Request, res: Response) => {
     const users = await this._usersUsecase.allUsers();
     res.json(users);
     return;
-  }
+  };
 
-  async getUserData(req: Request, res: Response) {
+  public getUserData = async (req: Request, res: Response) => {
     const user = await this._usersUsecase.getUser(req.params.userId);
     res.json(user);
     return;
-  }
+  };
 
-  async login(req: Request, res: Response) {
+  public login = async (req: Request, res: Response) => {
     const users = await this._usersUsecase.genJWTToken(
       req.body.name,
       req.body.password
@@ -31,9 +31,9 @@ export class UsersController {
     };
     res.json(resBody);
     return;
-  }
+  };
 
-  async register(req: Request, res: Response) {
+  public register = async (req: Request, res: Response) => {
     const user = await this._usersUsecase.createUser(
       req.body.name,
       req.body.password
@@ -43,5 +43,5 @@ export class UsersController {
     };
     res.json(resBody);
     return;
-  }
+  };
 }
